Refactor docker module out of orbital_agent Created new crate orbital_exec_runtime and moved docker module there
