+++ cli/orb/src/subcommand/Cargo.toml
@@ -12,6 +12,7 @@ git2 = "0.13"
 git-url-parse = "0.2"
 log = "0.4"
 orbital_agent = { path = "../../../../lib/orbital_agent" }
+orbital_exec_runtime = { path = "../../../../lib/exec_runtime" }
 orbital_config_parser = { path = "../../../../lib/config_parser" }
 orbital_database = { path = "../../../../models/orbital_database" }
 orbital_git_meta = { path = "../../../../lib/git_meta" }
