+++ cli/orb/src/subcommand/developer/docker/exec.rs
@@ -1,7 +1,7 @@
 use crate::{developer::docker::SubcommandOption, GlobalOption, SubcommandError};
-use orbital_agent::docker;
 use anyhow::Result;
 use log::debug;
+use orbital_exec_runtime::docker;
 use structopt::StructOpt;
 
 #[derive(Debug, StructOpt, Clone)]
