+++ lib/hashicorp_stack/vault.rs
