+++ lib/orbital_agent/build_engine.rs
@@ -1,12 +1,12 @@
-use crate::docker::{self, OrbitalContainerSpec};
 use crate::AgentRuntimeError;
 use anyhow::Result;
 use config_parser;
 use git_meta;
 use log::{debug, info};
 use mktemp;
-use std::path::Path;
+use orbital_exec_runtime::docker::{self, OrbitalContainerSpec};
 use state_machine;
+use std::path::Path;
 
 use serde_json::value::Value;
 use tokio::sync::mpsc;
