+++ models/orbital_database/postgres/client.rs
@@ -7,10 +7,10 @@ use crate::postgres::schema::{
     build_stage, build_summary, build_target, org, repo, secret, JobState, JobTrigger, SecretType,
 };
 use crate::postgres::secret::{NewSecret, Secret};
-use hashicorp_stack::orb_vault_path;
 use anyhow::{anyhow, Result};
 use diesel::pg::PgConnection;
 use diesel::prelude::*;
+use hashicorp_stack::orb_vault_path;
 use log::debug;
 use std::env;
 //use orbital_headers::orbital_types;
