+++ services/Cargo.toml
@@ -11,6 +11,7 @@ futures = "0.3"
 log = "0.4"
 mktemp = "0.4"
 orbital_agent = { path = "../lib/orbital_agent" }
+orbital_exec_runtime = { path = "../lib/exec_runtime" }
 orbital_config_parser = { path = "../lib/config_parser" }
 orbital_database = { path = "../models/orbital_database" }
 orbital_git_meta = { path = "../lib/git_meta" }
