+++ services/code_service.rs
@@ -18,8 +18,8 @@ use tonic::{Request, Response, Status};
 
 use super::OrbitalApi;
 
-use orbital_agent::build_engine;
 use log::{debug, info};
+use orbital_agent::build_engine;
 use orbital_database::postgres;
 
 use serde_json::json;
