+++ services/secret_service.rs
@@ -6,8 +6,8 @@ use tonic::{Request, Response, Status};
 
 use super::OrbitalApi;
 
-use orbital_agent::vault;
 use log::debug;
+use orbital_agent::vault;
 use orbital_database::postgres;
 use orbital_database::postgres::schema::SecretType;
